@@ -7,6 +7,7 @@
     "build": "gulp build",
     "start": "gulp"
   },
+	"type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/AntonOshurek/Unicorm_page.git"
@@ -16,12 +17,12 @@
   ],
   "author": "Anton Oshurek",
   "browserslist": [
-    "last 2 versions",
+    "last 10 versions",
     "not dead",
     "not ie <= 11"
   ],
   "engines": {
-    "node": "^12 || ^14"
+    "node": ">=16"
   },
   "license": "ISC",
   "bugs": {
@@ -29,27 +30,26 @@
   },
   "homepage": "https://github.com/AntonOshurek/Unicorm_page#readme",
   "devDependencies": {
-    "@babel/cli": "^7.15.7",
-    "@babel/core": "^7.15.5",
-    "@babel/preset-env": "^7.15.6",
-    "autoprefixer": "10.2.5",
-    "babel-loader": "^8.2.2",
-    "browser-sync": "2.26.14",
-    "core-js": "^2.6.12",
-    "del": "^6.0.0",
-    "gulp": "4.0.2",
+    "@babel/cli": "^7.20.7",
+    "@babel/core": "^7.20.7",
+    "@babel/preset-env": "^7.20.2",
+    "autoprefixer": "10.4.13",
+    "babel-loader": "^9.1.0",
+    "browser-sync": "^2.27.11",
+    "core-js": "^3.27.0",
+    "del": "^7.0.0",
+    "gulp": "^4.0.2",
     "gulp-htmlmin": "^5.0.1",
-    "gulp-less": "4.0.1",
-    "gulp-libsquoosh": "^1.0.9",
+    "gulp-less": "5.0.0",
+    "gulp-libsquoosh": "^1.1.2",
     "gulp-plumber": "1.2.1",
-    "gulp-postcss": "9.0.0",
+    "gulp-postcss": "9.0.1",
     "gulp-rename": "^2.0.0",
     "gulp-sourcemaps": "3.0.0",
     "gulp-webp": "^4.0.1",
-    "postcss": "8.2.8",
-    "postcss-csso": "^5.0.1",
-    "webpack": "^5.53.0",
-    "webpack-cli": "^4.8.0",
+    "postcss": "^8.4.20",
+    "postcss-csso": "^6.0.1",
+    "webpack": "^5.75.0",
     "webpack-stream": "^7.0.0"
   },
   "dependencies": {
